Restructure test infrastructure types
rename Infrastructure to TestInfrastructure
