+++ tests/.editorconfig
@@ -1,7 +1,6 @@
 # C# files
 [**/*.cs]
 
-dotnet_diagnostic.CA1051.severity = suggestion # Do not declare visible instance fields
 dotnet_diagnostic.CA1859.severity = suggestion # Use concrete types when possible for improved performance
 dotnet_diagnostic.CA1861.severity = suggestion # Avoid constant arrays as arguments
 dotnet_diagnostic.CA2201.severity = suggestion # Do not raise reserved exception types
