+++ tests/LibUsbSharp.Tests/Given_no_USB_device.cs
@@ -1,5 +1,5 @@
 using FluentAssertions;
-using LibUsbSharp.Tests.Infrastructure;
+using LibUsbSharp.Tests.TestInfrastructure;
 using Microsoft.Extensions.Logging;
 using Xunit;
 using Xunit.Abstractions;
